Implemented forward algorithm on emission tensor
The a posteriori evidence of the hidden variable can now be calculated with the recursive forward algorithm for a given evidence sequence.
Updated HMM class to use vectors for states since these allow for direct access of tensor fields with indices in vectors.
Updated state generation algortihms to output vectors instead of ITensors.